BJS2-67674 Доработки
